Refactoring: Authoriztion Components
