+++ src/components/common/authorization/Authorized.tsx
@@ -6,21 +6,21 @@ import { useEffect } from 'react';
 interface Props {
   fallback?: React.ReactNode;
   children: React.ReactNode;
-  homeUrl: string;
+  signInUrl: string;
 }
 
-/** children 하위 컴포넌트는 session이 없음을 보장 */
-export default function UnAuthorized({ children, fallback, homeUrl }: Props) {
+/** children 하위 컴포넌트는 session이 있음을 보장 */
+export function Authorized({ children, fallback, signInUrl }: Props) {
   const session = useSession();
   const router = useRouter();
 
   useEffect(() => {
-    if (session.status === 'authenticated') {
-      router.replace(homeUrl);
+    if (session.status === 'unauthenticated') {
+      router.replace(signInUrl);
     }
   }, [session.status]);
 
   if (session.status === 'loading') return <>{fallback}</>;
-  if (session.status === 'authenticated') return null;
+  if (session.status === 'unauthenticated') return null;
   return <>{children}</>;
 }
